pyproject.toml
[tool.poetry]
name = "food-truck-finder"
version = "0.1.0"
description = ""
authors = ["RustamGarayev <rustam.qarayev02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
geoip2 = "^4.7.0"
dj-database-url = "^2.1.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
